major changes in code structure
