+++ core/src/main/java/dev/keva/core/command/impl/hash/HExists.java
@@ -11,17 +11,18 @@ import dev.keva.store.KevaDatabase;
 @Component
 @CommandImpl("hexists")
 @ParamLength(2)
-public class HExists {
+public class HExists extends HashBase {
     private final KevaDatabase database;
 
     @Autowired
     public HExists(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public IntegerReply execute(byte[] key, byte[] field) {
-        byte[] got = database.hget(key, field);
+        byte[] got = this.get(key, field);
         return got == null ? new IntegerReply(0) : new IntegerReply(1);
     }
 }
