+++ core/src/main/java/dev/keva/core/command/impl/hash/HGet.java
@@ -11,17 +11,18 @@ import dev.keva.store.KevaDatabase;
 @Component
 @CommandImpl("hget")
 @ParamLength(type = ParamLength.Type.AT_LEAST, value = 2)
-public class HGet {
+public class HGet extends HashBase {
     private final KevaDatabase database;
 
     @Autowired
     public HGet(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public BulkReply execute(byte[] key, byte[] field) {
-        byte[] got = database.hget(key, field);
+        byte[] got = this.get(key, field);
         return got == null ? BulkReply.NIL_REPLY : new BulkReply(got);
     }
 }
