+++ core/src/main/java/dev/keva/core/command/impl/hash/HGetAll.java
@@ -12,17 +12,18 @@ import dev.keva.store.KevaDatabase;
 @Component
 @CommandImpl("hgetall")
 @ParamLength(1)
-public class HGetAll {
+public class HGetAll extends HashBase {
     private final KevaDatabase database;
 
     @Autowired
     public HGetAll(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public MultiBulkReply execute(byte[] key) {
-        byte[][] got = database.hgetAll(key);
+        byte[][] got = this.getAll(key);
         BulkReply[] replies = new BulkReply[got.length];
         for (int i = 0; i < got.length; i++) {
             replies[i] = new BulkReply(got[i]);
