+++ core/src/main/java/dev/keva/core/command/impl/hash/HKeys.java
@@ -8,23 +8,35 @@ import dev.keva.ioc.annotation.Component;
 import dev.keva.protocol.resp.reply.BulkReply;
 import dev.keva.protocol.resp.reply.MultiBulkReply;
 import dev.keva.store.KevaDatabase;
+import dev.keva.util.hashbytes.BytesKey;
+import dev.keva.util.hashbytes.BytesValue;
+import org.apache.commons.lang3.SerializationUtils;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @Component
 @CommandImpl("hkeys")
 @ParamLength(1)
-public class HKeys {
+public class HKeys extends HashBase {
     private final KevaDatabase database;
 
     @Autowired
     public HKeys(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public MultiBulkReply execute(byte[] key) {
-        byte[][] got = database.hkeys(key);
+        HashMap<BytesKey, BytesValue> map = this.getMap(key);
+        byte[][] got = new byte[map.size()][];
+        int i = 0;
+        for (Map.Entry<BytesKey, BytesValue> entry : map.entrySet()) {
+            got[i++] = entry.getKey().getBytes();
+        }
         BulkReply[] replies = new BulkReply[got.length];
-        for (int i = 0; i < got.length; i++) {
+        for (i = 0; i < got.length; i++) {
             replies[i] = new BulkReply(got[i]);
         }
         return new MultiBulkReply(replies);
