+++ core/src/main/java/dev/keva/core/command/impl/hash/HStrLen.java
@@ -11,17 +11,18 @@ import dev.keva.store.KevaDatabase;
 @Component
 @CommandImpl("hstrlen")
 @ParamLength(2)
-public class HStrLen {
+public class HStrLen extends HashBase {
     private final KevaDatabase database;
 
     @Autowired
     public HStrLen(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public IntegerReply execute(byte[] key, byte[] field) {
-        byte[] got = database.hget(key, field);
+        byte[] got = this.get(key, field);
         return got == null ? new IntegerReply(0) : new IntegerReply(got.length);
     }
 }
