+++ core/src/main/java/dev/keva/core/command/impl/list/LRem.java
@@ -13,17 +13,18 @@ import dev.keva.store.KevaDatabase;
 @CommandImpl("lrem")
 @ParamLength(3)
 @Mutate
-public class LRem {
+public class LRem extends ListBase {
     private final KevaDatabase database;
 
     @Autowired
     public LRem(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public IntegerReply execute(byte[] key, byte[] count, byte[] value) {
-        int result = database.lrem(key, Integer.parseInt(new String(count)), value);
+        int result = this.remove(key, Integer.parseInt(new String(count)), value);
         return new IntegerReply(result);
     }
 }
