+++ core/src/main/java/dev/keva/core/command/impl/list/LSet.java
@@ -13,17 +13,18 @@ import dev.keva.store.KevaDatabase;
 @CommandImpl("lset")
 @ParamLength(3)
 @Mutate
-public class LSet {
+public class LSet extends ListBase {
     private final KevaDatabase database;
 
     @Autowired
     public LSet(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public StatusReply execute(byte[] key, byte[] index, byte[] value) {
-        database.lset(key, Integer.parseInt(new String(index)), value);
+        this.set(key, Integer.parseInt(new String(index)), value);
         return StatusReply.OK;
     }
 }
