+++ core/src/main/java/dev/keva/core/command/impl/list/RPush.java
@@ -15,17 +15,18 @@ import java.util.Arrays;
 @CommandImpl("rpush")
 @ParamLength(type = ParamLength.Type.AT_LEAST, value = 2)
 @Mutate
-public class RPush {
+public class RPush extends ListBase {
     private final KevaDatabase database;
 
     @Autowired
     public RPush(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public IntegerReply execute(byte[]... params) {
-        int count = database.rpush(params[0], Arrays.copyOfRange(params, 1, params.length));
+        int count = this.rpush(params[0], Arrays.copyOfRange(params, 1, params.length));
         return new IntegerReply(count);
     }
 }
