+++ core/src/main/java/dev/keva/core/command/impl/set/SIsMember.java
@@ -11,17 +11,18 @@ import dev.keva.store.KevaDatabase;
 @Component
 @CommandImpl("sismember")
 @ParamLength(2)
-public class SIsMember {
+public class SIsMember extends SetBase {
     private final KevaDatabase database;
 
     @Autowired
     public SIsMember(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public IntegerReply execute(byte[] key, byte[] value) {
-        boolean isMember = database.sismember(key, value);
+        boolean isMember = this.isMember(key, value);
         return new IntegerReply(isMember ? 1 : 0);
     }
 }
