+++ core/src/main/java/dev/keva/core/command/impl/set/SRem.java
@@ -15,17 +15,18 @@ import java.util.Arrays;
 @CommandImpl("srem")
 @ParamLength(type = ParamLength.Type.AT_LEAST, value = 2)
 @Mutate
-public class SRem {
+public class SRem extends SetBase {
     private final KevaDatabase database;
 
     @Autowired
     public SRem(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public IntegerReply execute(byte[][] params) {
-        int removed = database.srem(params[0], Arrays.copyOfRange(params, 1, params.length));
+        int removed = this.remove(params[0], Arrays.copyOfRange(params, 1, params.length));
         return new IntegerReply(removed);
     }
 }
