+++ core/src/main/java/dev/keva/core/command/impl/set/SUnion.java
@@ -12,17 +12,18 @@ import dev.keva.store.KevaDatabase;
 @Component
 @CommandImpl("sunion")
 @ParamLength(2)
-public class SUnion {
+public class SUnion extends SetBase {
     private final KevaDatabase database;
 
     @Autowired
     public SUnion(KevaDatabase database) {
+        super(database);
         this.database = database;
     }
 
     @Execute
     public MultiBulkReply execute(byte[]... keys) {
-        byte[][] diff = database.sunion(keys);
+        byte[][] diff = this.union(keys);
         BulkReply[] replies = new BulkReply[diff.length];
         for (int i = 0; i < diff.length; i++) {
             replies[i] = new BulkReply(diff[i]);
