+++ core/src/main/java/dev/keva/core/command/impl/string/Decrby.java
@@ -29,7 +29,14 @@ public class Decrby {
         long amount = Long.parseLong(new String(decrBy, StandardCharsets.UTF_8));
         byte[] newVal;
         try {
-            newVal = database.decrby(key, amount);
+            newVal = database.compute(key, (k, oldVal) -> {
+                long curVal = 0L;
+                if (oldVal != null) {
+                    curVal = Long.parseLong(new String(oldVal, StandardCharsets.UTF_8));
+                }
+                curVal = curVal - amount;
+                return Long.toString(curVal).getBytes(StandardCharsets.UTF_8);
+            });
         } catch (NumberFormatException ex) {
             throw new CommandException("value is not an integer or out of range");
         }
