+++ core/src/main/java/dev/keva/core/command/impl/string/Incrby.java
@@ -31,7 +31,14 @@ public class Incrby {
         long amount = Long.parseLong(new String(incrBy, StandardCharsets.UTF_8));
         byte[] newVal;
         try {
-            newVal = database.incrBy(key, amount);
+            newVal = database.compute(key, (k, oldVal) -> {
+                long curVal = 0L;
+                if (oldVal != null) {
+                    curVal = Long.parseLong(new String(oldVal, StandardCharsets.UTF_8));
+                }
+                curVal = curVal + amount;
+                return Long.toString(curVal).getBytes(StandardCharsets.UTF_8);
+            });
         } catch (NumberFormatException ex) {
             throw new CommandException("Failed to parse integer from value stored");
         }
