+++ core/src/main/java/dev/keva/core/command/impl/string/MGet.java
@@ -25,7 +25,12 @@ public class MGet {
     @Execute
     public MultiBulkReply execute(byte[]... keys) {
         BulkReply[] replies = new BulkReply[keys.length];
-        byte[][] mget = database.mget(keys);
+        byte[][] mget = new byte[keys.length][];
+        for (int i = 0; i < keys.length; i++) {
+            byte[] key = keys[i];
+            byte[] got = database.get(key);
+            mget[i] = got;
+        }
         for (int i = 0; i < mget.length; i++) {
             replies[i] = mget[i] == null ? BulkReply.NIL_REPLY : new BulkReply(mget[i]);
         }
