+++ core/src/main/java/dev/keva/core/command/impl/string/MSet.java
@@ -27,7 +27,9 @@ public class MSet {
         if (keys.length % 2 != 0) {
             throw new CommandException("Wrong number of arguments for MSET");
         }
-        database.mset(keys);
+        for (int i = 0; i < keys.length; i += 2) {
+            database.put(keys[i], keys[i + 1]);
+        }
         return StatusReply.OK;
     }
 }
