+++ core/src/main/java/dev/keva/core/command/impl/string/SetEX.java
@@ -6,12 +6,8 @@ import dev.keva.core.command.annotation.ParamLength;
 import dev.keva.core.exception.CommandException;
 import dev.keva.ioc.annotation.Autowired;
 import dev.keva.ioc.annotation.Component;
-import dev.keva.protocol.resp.reply.BulkReply;
-import dev.keva.protocol.resp.reply.Reply;
 import dev.keva.protocol.resp.reply.StatusReply;
 import dev.keva.store.KevaDatabase;
-
-import java.nio.charset.StandardCharsets;
 
 @Component
 @CommandImpl("setex")
@@ -26,7 +22,6 @@ public class SetEX {
 
     @Execute
     public StatusReply execute(byte[][] params) {
-        String key = new String(params[0], StandardCharsets.UTF_8);
         long seconds;
         try {
             seconds = Long.parseLong(new String(params[1]));
