+++ core/src/main/java/dev/keva/core/command/impl/string/SetNX.java
@@ -3,15 +3,11 @@ package dev.keva.core.command.impl.string;
 import dev.keva.core.command.annotation.CommandImpl;
 import dev.keva.core.command.annotation.Execute;
 import dev.keva.core.command.annotation.ParamLength;
-import dev.keva.core.exception.CommandException;
 import dev.keva.ioc.annotation.Autowired;
 import dev.keva.ioc.annotation.Component;
 import dev.keva.protocol.resp.reply.IntegerReply;
 import dev.keva.store.KevaDatabase;
 
-import java.util.Arrays;
-
-import static dev.keva.core.command.annotation.ParamLength.Type.AT_LEAST;
 import static dev.keva.core.command.annotation.ParamLength.Type.EXACT;
 
 @Component
