+++ core/src/main/java/dev/keva/core/command/impl/string/SetRange.java
@@ -9,6 +9,8 @@ import dev.keva.ioc.annotation.Component;
 import dev.keva.protocol.resp.reply.IntegerReply;
 import dev.keva.store.KevaDatabase;
 
+import java.nio.charset.StandardCharsets;
+
 import static dev.keva.core.command.annotation.ParamLength.Type.EXACT;
 
 @Component
@@ -25,6 +27,20 @@ public class SetRange {
 
     @Execute
     public IntegerReply execute(byte[] key, byte[] offset, byte[] val) {
-        return new IntegerReply(database.setrange(key, offset, val));
+        int offsetPosition = Integer.parseInt(new String(offset, StandardCharsets.UTF_8));
+        byte[] oldVal = database.get(key);
+        int newValLength = oldVal == null ? offsetPosition + val.length : Math.max(offsetPosition + val.length, oldVal.length);
+        byte[] newVal = new byte[newValLength];
+        for (int i = 0; i < newValLength; i++) {
+            if (i >= offsetPosition && i < offsetPosition + val.length) {
+                newVal[i] = val[i - offsetPosition];
+            } else if (oldVal != null && i < oldVal.length) {
+                newVal[i] = oldVal[i];
+            } else {
+                newVal[i] = 0b0;
+            }
+        }
+        database.put(key, newVal);
+        return new IntegerReply(newValLength);
     }
 }
